qsafe-demo/requirements.txt
flask==2.3.3
Flask-SocketIO==5.3.6
pycryptodome==3.19.0
colorama==0.4.6
requests==2.31.0
websocket-client==1.6.4